% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking.R
\name{BlockBySubstr}
\alias{BlockBySubstr}
\title{Block by substrings}
\usage{
BlockBySubstr(records, var.names, n.chars = NULL)
}
\arguments{
\item{records}{Data Frame containing records to be linked}

\item{var.names}{Vector of strings of variables you want to block by}

\item{n.chars}{Vector of number of the characters you want to compare, respective to var.names}
}
\value{
A list with the elements
\item{blocks}{vector of strings corresponding to the blocks}
\item{factors}{vector of integers corresponding to the blocks}
}
\description{
Block by substrings of any length of multiple variables
}
\examples{
BlockBySubstr(iris, "Species") #identifies 2 blocks
BlockBySubstr(iris, "Species", 2) #identifies 3 blocks
BlockBySubstr(iris, c("Species", "Sepal.Length"), c(2,1)) #identifies 3 blocks
}

