% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking.R
\name{BlockBySubstrMV}
\alias{BlockBySubstrMV}
\title{Block by substring}
\usage{
BlockBySubstrMV(records, var1.name, n1.chars = 1, var2.name = NULL,
  n2.chars = 1, var3.name = NULL, n3.chars = 1, var4.name = NULL,
  n4.chars = 1)
}
\arguments{
\item{records}{Data Frame containing records to be linked}

\item{var.name}{String of variable you want to block by}

\item{n.chars}{Number of characters you want to compare, default is 1}
}
\value{
A printed vector of integers coresponding to the block
}
\description{
Block by substrings of up to 4 variables
}
\examples{
BlockByInitials(iris, "Species") #identifies 2 blocks
BlockByInitials(iris, "Species", 2) #identifies 3 blocks
}

