% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking.R
\name{BlockByVariable}
\alias{BlockByVariable}
\title{Block by variable}
\usage{
BlockByVariable(records, var.name)
}
\arguments{
\item{records}{Data Frame containing records to be linked}

\item{var.name}{String of variable you want to block by}
}
\value{
A vector of integers coresponding to the block assignment
}
\description{
Block a dataset by the factor levels of one variable
}
\examples{
BlockByVariable(iris, "Species")
}

