% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking.R
\name{BlockInPasses}
\alias{BlockInPasses}
\title{Block a record linkage dataset in passes}
\usage{
BlockInPasses(records, pass.structure)
}
\arguments{
\item{records}{a data frame containing the records to be matched}

\item{pass.structure}{a list containing a matrix for each pass where the first column of the matrix contains the variables to block on and the second column contains the number of characters to use (NA will use the entire variable)}
}
\value{
A data frame containing the ids of records we will compare and the blocking scheme used to choose them
}
\description{
Block a record linkage dataset in passes of different blocking schemes
}
\examples{
BlockInPasses(RLdata10000, list(matrix(c("fname_c1", "lname_c1", NA, NA), ncol = 2),
                                matrix(c("fname_c1", "by", NA, NA), ncol = 2),
                                matrix(c("fname_c1", "lname_c1", 3, 4), ncol = 2),
                                matrix(c("lname_c1", NA), ncol = 2)))

BlockInPasses(RLdata500, list(matrix(c("fname_c1", "lname_c1", "by", 1, 2, NA), ncol = 2)))

}

