% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking.R
\name{BlockRlData}
\alias{BlockRlData}
\title{Block a record linkage dataset}
\usage{
BlockRlData(RLdata, var.names, n.chars = NULL, unique.ids = NULL)
}
\arguments{
\item{RLdata}{a data frame containing the records to be matched}

\item{var.names}{a vector of strings containing the variable names you want to block by}

\item{n.chars}{a vector of integers corresponding to the number of the characters you want to compare in each variable of var.names}

\item{unique.ids}{a vector containing the true unique identifiers of the records in RLdata. It should be of length nrow(RLdata)}
}
\value{
A list containing blocking information and the blocked data and ids
\item{BlockInfo}{a list of blocking information: blocks, factors, reduction.ratio}
\item{DataSplit}{a list of datasets corresponding to each block}
\item{IdSplit}{a list of vectors containing the unique ids corresponding to each block}
}
\description{
Block a record linkage dataset by substrings of the variables in the dataset
}
\examples{
BlockBySubstr(iris, "Species") #identifies 2 blocks
BlockBySubstr(iris, "Species", 2) #identifies 3 blocks
BlockBySubstr(iris, c("Species", "Sepal.Length"), c(2,1)) #identifies 3 blocks

}

