% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{BuildATrainingDataset}
\alias{BuildATrainingDataset}
\title{Build a training dataset}
\usage{
BuildATrainingDataset(RLdata, n.pairs.to.test, variables.to.match = NULL,
  string.comparators = NULL, current.record.ids = NULL,
  standardized.variables = NULL, seed = NULL)
}
\arguments{
\item{RLdata}{a data frame containing the records to be matched}

\item{n.pairs.to.test}{an integer corresponding to the number of pairs of records the user wants to test}

\item{variables.to.match}{a vector of strings containing the variables of interest for this linkage. Default is all variables in RLdata. Can repeat variables to use different comparators on same variable.}

\item{string.comparators}{a vector of strings containing the comparator to be used for each variable. Default is jarowinkler. This should be same length as variables.to.match.}

\item{current.record.ids}{a vector of strings corresponding to the variable names that will contain the pairwise combinations of records. The default (which is produced using CompareUniqueCombinations) is c("CurrentRecord1", "CurrentRecord2").}

\item{standardized.variables}{a vector of strings containing the names of all standardized variables. The comparison values for these variables will be averaged. The default is all factor variables in RLdata.}
}
\value{
A list with the elements
\item{comparisons}{a data frame containing the comparisons of RLdata}
\item{tested.comparisons}{a data frame containing the comparison values of records the user tested}
\item{tested.data}{a data frame containing the original data of records the user tested}
}
\description{
Build a training dataset from user input about whether records in the dataset match
}

