% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{BuildATrainingDatasetAuto}
\alias{BuildATrainingDatasetAuto}
\title{Build a training dataset automatically}
\usage{
BuildATrainingDatasetAuto(RLdata, block.comparisons, standardized.variables,
  model.formula, n.pairs.to.test = NULL, n.per.stage = NULL,
  n.initial.stage = NULL, initial.matching.scheme = NULL,
  record.id = NULL, user.accuracy = NULL, uncertainty.param = NULL,
  seed = NULL)
}
\arguments{
\item{RLdata}{a data frame containing the records to be matched}

\item{standardized.variables}{a vector of strings containing the names of all standardized variables. The comparison values for these variables will be averaged. The default is all factor variables in RLdata.}

\item{n.pairs.to.test}{an integer corresponding to the number of pairs of records the user wants to test}

\item{record.ids}{a vector of strings corresponding to the variable names that will contain the pairwise combinations of records. The default (which is produced using CompareUniqueCombinations) is c("CurrentRecord1", "CurrentRecord2").}
}
\value{
A list with the elements
\item{comparisons}{a data frame containing the comparisons of RLdata}
\item{tested.comparisons}{a data frame containing the comparison values of records the user tested}
\item{untested.comparisons}{a data frame containing the comparison values of records the user did not test}
}
\description{
Build a training dataset from user input about whether records in the dataset match
}

