% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{BuildTrainingClusterFull}
\alias{BuildTrainingClusterFull}
\title{Get unique ids actively}
\usage{
BuildTrainingClusterFull(RLdata, RLdata.blocked, block.comparisons,
  standardized.variables, model.formula, n.pairs.to.test = NULL,
  n.per.stage = NULL, n.initial.stage = NULL,
  initial.matching.scheme = NULL, record.id = NULL,
  uncertainty.param = NULL, seed = NULL, cut.threshold = NULL)
}
\description{
Get unique ids actively
}

