% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparing.R
\name{CompareUniqueCombinations}
\alias{CompareUniqueCombinations}
\title{Compare Unique Combinations of Records}
\usage{
CompareUniqueCombinations(RLdata, unique.ids = NULL,
  variables.to.match = NULL, string.comparators = NULL,
  record.ids.to.keep = NULL)
}
\arguments{
\item{RLdata}{Dataset containing records to be matched}

\item{unique.ids}{True unique identification vector of length nrow(RLdata)}

\item{variables.to.match}{Vector of strings containing the variables of interest for this linkage. Default is all variables in RLdata. Can repeat variables to use different comparators on same variable.}

\item{string.comparators}{Vector of strings containing the comparators wanted for each variable. Default is jarowinkler. Should be same length as variables.to.match.}
}
\value{
A matrix of comparator values of each combination of records for the variables of interest. Also includes binary matches and the combinations of records.
}
\description{
Match records from a dataset using any comparison function on any set of variables
}

