% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparing.R
\name{ContToBin}
\alias{ContToBin}
\title{Turn all continuous variables into binary}
\usage{
ContToBin(dat, cat.vars, cut.offs, inequals, keep.orig = FALSE)
}
\value{
A data frame
}
\description{
Turn all continuous variables into binary
}

