% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fellegi.R
\name{FellegiSunter}
\alias{FellegiSunter}
\title{Fellegi Sunter}
\usage{
FellegiSunter(full.data, initial.m = NULL, initial.u = NULL,
  initial.p = NULL, converge.at = NULL, verbose = FALSE)
}
\arguments{
\item{full.data}{}

\item{initial.m}{}

\item{initial.u}{}

\item{initial.p}{}

\item{converge.at}{}

\item{verbose}{}
}
\value{
Blah blah
}
\description{
Fellegi Sunter
}

