% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{FSActiveFull}
\alias{FSActiveFull}
\title{Get unique ids actively}
\usage{
FSActiveFull(RLdata, RLdata.blocked, block.comparisons, cols.for.fs,
  n.pairs.to.test = NULL, n.per.stage = NULL, n.initial.stage = NULL,
  record.id = NULL, uncertainty.param = NULL, seed = NULL,
  cut.threshold = NULL, save.self.matched = FALSE)
}
\description{
make sure n initial stage is > 200
#' #' @param record.ids a vector of strings corresponding to the variable names that will contain the pairwise combinations of records in the data. The default (which is produced using CompareUniqueCombinations) is c("CurrentRecord1", "CurrentRecord2").
I think we want it to be PreBlockRecord if we've blocked?
}

