% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparing.R
\name{GetPairwiseMatchesFromIDs}
\alias{GetPairwiseMatchesFromIDs}
\title{Get pairwise matches from IDs}
\usage{
GetPairwiseMatchesFromIDs(combinations.of.original.data, ids.from.original.data)
}
\arguments{
\item{combinations.of.original.data}{The combinations of your original data that you want to extract pairwise matches for}

\item{ids.from.original.data}{The vector of unique IDs from original data containing match information}
}
\value{
A vector of binary non-match/match values for each combination
}
\description{
Take a vector of unique IDs from original data set and convert to binary matches for the comparisons of each entity in original dataset
}
\examples{
# use packages RecordLinkage and gtools
# return vector of nrow(our.combinations)
our.combinations <- combinations(nrow(RLdata500), 2)
GetPairwiseMatchesFromIDs(our.combinations, identity.RLdata500)

}

