% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{IsItAMatch}
\alias{IsItAMatch}
\title{Ask the user if two records match}
\usage{
IsItAMatch(record1, record2)
}
\arguments{
\item{record1}{a row of data corresponding to the first record we are interested in}

\item{record2}{a row of data corresponding to the second record we are interested in}
}
\value{
y/n if the two records match
}
\description{
Prompt the user to input yes or no (y/n) if two records match
}
\examples{
IsItAMatch(RLdata500[1, ], RLdata[2, ])

}

