% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparing.R
\name{MergeAllBlocks}
\alias{MergeAllBlocks}
\title{Merge (rbind) all datasets in a list}
\usage{
MergeAllBlocks(block.lists)
}
\arguments{
\item{block.lists}{a list of datasets}
}
\value{
A data frame with all elements of block.lists merged together
}
\description{
Merge (rbind) all datasets in a list
}

