% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{ProbItsAMatch}
\alias{ProbItsAMatch}
\title{Ask the user for the probability two records match}
\usage{
ProbItsAMatch(record1, record2)
}
\arguments{
\item{record1}{a row of data corresponding to the first record we are interested in}

\item{record2}{a row of data corresponding to the second record we are interested in}
}
\value{
The probability that the two records match
}
\description{
Prompt the user to input the probability that two records match
}
\examples{
ProbItsAMatch(RLdata500[1, ], RLdata[2, ])
}

